+++ src/clang/incl/comm.h
@@ -152,7 +152,7 @@ typedef struct
 #endif
 
 typedef int64_t (*key_cb_t)(const void *data);
-typedef int (*cmp_cb_t)(void *data1, void *data2);
+typedef int64_t (*cmp_cb_t)(void *data1, void *data2);
 typedef int (*trav_cb_t)(void *data, void *args);
 typedef bool (*find_cb_t)(void *data, void *args);
 typedef void (*print_cb_t)(void *data);
